backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.2.18
django-cors-headers==3.14.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
docopt==0.6.2
idna==3.4
mongoengine==0.27.0
pipreqs==0.4.11
pymongo==3.12.1
python-dotenv==0.21.1
pytz==2023.2
requests==2.28.2
sqlparse==0.2.4
typing_extensions==4.5.0
urllib3==1.26.15
yarg==0.1.9
